(GH-141) Fix Icons spin when changing visibility (UWP)
When changing the visibility of an icon (from Collapsed to Visible) in an UWP project, the icon starts spinning.
This is now fixed after a closer look. In UWP there is no Coerce value callback like in WPF to trigger the spin property changed callback and react to the visibility or opacity. Therefore this has to be done by a toggle mechanism, which was implemented wrong.